requirements.txt
python-telegram-bot
Flask
pillow
